refactor(#86): Contract, ViewModel 이름 변경 - 공지 및 투표의 리스트를 보여주던 것 관련 Contract 및 ViewModel을 ~~List~~로 변경
